do some cacheing
